refactor: Move saved list use cases and tests to dedicated package
