+++ remote_datasource/src/test/java/com/baghdad/remoteDataSource/mapper/savedList/SavedListsResponseMapperTest.kt
@@ -37,8 +37,6 @@ class SavedListsResponseMapperTest {
     @Test
     fun `toPagedSavedListsDtos should return empty list when results is null`() {
         // Given
-
-
         val response = UserListsResponse(
             page = 1,
             results = null,
